Add isHardReboot() code to check if hard reboot is required
The default reboot-interface behaviour is to attempt a soft power
off, and if this fails, revert to a hard power off (PR #294). For high
availability use-cases we require the ability to immediately power-off
a node. This PR attempts to address that requirement and is a WIP given
that other changes are going to be required to the provisioner API and
the CAPBM actuator to actually enact these changes.

Also see: https://bugzilla.redhat.com/show_bug.cgi?id=1927678