fixups
Change isHardReboot to getRebootMode so the controller code can just
pass the result to the provisioner.

Swap the order of arguments to assert.Equal() to place the expected
value in the right spot.

Fix a couple of tests that were expecting an empty string for the
reboot mode, since that value should never be an empty string.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>